Add placeholder "Roster" and "About" tabs
Add two additional tabs to the main navigation bar of the webapp. Both
tabs do not currently have any content.

The Roster tab will be the first tab selected after login, and will
display the class roster for a given section.

The About tab will always be enabled. A future revision may make this a
link rather than a tab.