Restructure login page to accept instructor email
The login page now asks for an instructor email and password, and no
longer requires an instructor id. The position of various fields have
also been changed.

This commit results in the HTML page no longer functioning with the
corresponding JavaScript page. This will be addressed in a future commit