Implement new login functionality
Use /login REST API call to retrieve instructor information, including
instructor ID.

This AJAX call is processed when the login button is pressed, and uses
the email field from the new revision of the HTML page.

Parsing of the response of /login is not guaranteed to be currently
correct, as the server side behavior of this functionality has not been
fully specified or implemented.

The index.html page has also been slightly modified to better separate
out the two types of fields, and to update the field name of email.