Implement new login tab; minor fixes
The Web Client now has two major states: pre-login and post-login.

Before login, there are only two tabs visible, Login and About. In the
login tab, the DB Info and login fields can be filled out. Upon clicking
the login button, the client will attempt to connect to the server. If
successful, it will move to the post-login state.

In the post-login state, the Login tab disappears, and the Roster,
Attendance, Grades, and Reports tabs appear in its place. The Attendance
tab is automatically selected. Also, a small section in the upper right
hand corner appears, which contains the instructor's full name, and an
arrow. Clicking on the arrow reveals a logout button, which returns the
client to the pre-login state.

The about tab is visible and enabled in both states.

Minor fixes:

Remove extra </td> tag

Trim email field before sending (minor since server also performs this)