Make staging table and function temporary; cleanup
The staging table and helper function are now placed in the pg_temp
schema, which follows the conventions of other import scripts, and
eliminates the need to drop them.

Add comment clarifying that the current working directory needs to be
the same as the location of the script.

Lines 64 and 65 are wrapped more appropriately