Replace default JS alert dialog, misc tweaks
The use of the alert() facility has been replaced with the use of a
generic modal from the Materialize framework. This generic modal can be
accessed by calling showAlert(htmlContent), where htmlContent is any
valid html text, which will be displayed in the body of the dialog.

The compact version of the attendance table has been made more compact
by altering the font size.

Prefix icons have been added to the email and password fields.

Location of code for displaying missing fields alert has been improved.