Remove hardcoded schema from createTables
Previously, the createTables script assumed that the tables were being populated into a schema named `gradebook`.

Tables are now created without schema qualification and are placed into the schema as determined by the SET LOCAL SCHEMA statement on L26 (currently set to `alpha`). This command can only be used within a transaction, and so the table creation now occurs within a single transaction (which is desireable for other purposes as well).