Remove `Gradebook.` schema qualification
This commit remove all occurences of the `Gradebook.` schema qualification from table and function names. This allows for an easier implementation of multi-tenance as well as being part of allowing installation in a ClassDB environment by a student.

A simple replace-all-in-files was performed, with one exception of "Gradebook." being the end of a sentence in a comment.