Update gradebookServer.js to close db connection on error
I have included two additional client.end() statements in the executeQuery function. These statements close the connection to the database when there is an error thrown by a sent query. This allows for multiple errors to be thrown without a buildup of unclosed connections, which might exceed the connection limit imposed by certain database applications, such as ClassDB.